modifying for get history detail at FoTransactionHeaderRepository.java FoTransactionService.java FoundationHistoryMapper.java InvoiceManagementController.java InvoiceManagementService.java enhance EmailService.java
