+++ src/main/java/bca/mbb/adaptor/FeignClientService.java
@@ -6,7 +6,11 @@ import bca.mbb.feign.ExternalCorporatelClient;
 import bca.mbb.feign.ExternalEmailClient;
 import bca.mbb.feign.UploadManagementClient;
 import bca.mbb.mbbcommonlib.exception.GeneralException;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
@@ -19,7 +23,7 @@ import java.lang.reflect.Method;
 @Service
 @RequiredArgsConstructor
 public class FeignClientService {
-
+    protected final Logger logger = LoggerFactory.getLogger(this.getClass());
     private final ExternalEmailClient externalEmailClients;
     private final ExternalCorporatelClient externalCorporatelClient;
     private final UploadManagementClient uploadManagementClient;
@@ -34,6 +38,7 @@ public class FeignClientService {
             Method method = externalClient.getClass().getDeclaredMethod(env.getProperty(coreApiEnum.getMethodName()), RequestClientDto.class);
             return method.invoke(externalClient, requestClient);
         } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
+            logger.info("exception when call:{} at time:{}",env.getProperty(coreApiEnum.getMethodName()), LocalDateTime.now());
             throw new GeneralException(e);
         }
     }
