+++ src/main/java/bca/mbb/enums/CoreApiEnum.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 
 @Getter
 public enum CoreApiEnum {
-    EMAIL_USER("class-name.core.api.external", "method-name.email.user"),
-    EMAIL_CORPORATE("class-name.core.api.external.corporate", "method-name.email.corporate"),
-    GET_GROUP("class-name.core.api.management", "method-name.group");
+    EMAIL_USER("ExternalClientService", "getEmailUser"),
+    EMAIL_CORPORATE("ExternalCorporateService", "getEmailCorporate"),
+    GET_GROUP("InvoiceManagementClientService", "getGroupId");
 
     private String className;
     private String methodName;
